https://jira.catena-x.net/browse/A1ODT-70 move graph to own component
