+++ devops-tooling/dashboard/src/components/Dashboard/Dashboard.scss
@@ -1,4 +1,3 @@
 .dashboard {
-  height: 100%;
   background-color: #777;
 }
