+++ devops-tooling/dashboard/src/components/NetworkGraph/NetworkGraph.scss
@@ -1,0 +1,3 @@
+.network{
+  height: 100%;
+}
